Remove duplicate Month and Weekday enums
Well, they were duplicately-exposed, but triplicately-defined.
