+++ core/core_bind.cpp
@@ -613,27 +613,6 @@ void OS::_bind_methods() {
 	BIND_ENUM_CONSTANT(RENDERING_DRIVER_VULKAN);
 	BIND_ENUM_CONSTANT(RENDERING_DRIVER_OPENGL3);
 
-	BIND_ENUM_CONSTANT(DAY_SUNDAY);
-	BIND_ENUM_CONSTANT(DAY_MONDAY);
-	BIND_ENUM_CONSTANT(DAY_TUESDAY);
-	BIND_ENUM_CONSTANT(DAY_WEDNESDAY);
-	BIND_ENUM_CONSTANT(DAY_THURSDAY);
-	BIND_ENUM_CONSTANT(DAY_FRIDAY);
-	BIND_ENUM_CONSTANT(DAY_SATURDAY);
-
-	BIND_ENUM_CONSTANT(MONTH_JANUARY);
-	BIND_ENUM_CONSTANT(MONTH_FEBRUARY);
-	BIND_ENUM_CONSTANT(MONTH_MARCH);
-	BIND_ENUM_CONSTANT(MONTH_APRIL);
-	BIND_ENUM_CONSTANT(MONTH_MAY);
-	BIND_ENUM_CONSTANT(MONTH_JUNE);
-	BIND_ENUM_CONSTANT(MONTH_JULY);
-	BIND_ENUM_CONSTANT(MONTH_AUGUST);
-	BIND_ENUM_CONSTANT(MONTH_SEPTEMBER);
-	BIND_ENUM_CONSTANT(MONTH_OCTOBER);
-	BIND_ENUM_CONSTANT(MONTH_NOVEMBER);
-	BIND_ENUM_CONSTANT(MONTH_DECEMBER);
-
 	BIND_ENUM_CONSTANT(SYSTEM_DIR_DESKTOP);
 	BIND_ENUM_CONSTANT(SYSTEM_DIR_DCIM);
 	BIND_ENUM_CONSTANT(SYSTEM_DIR_DOCUMENTS);
