+++ core/os/os.h
@@ -34,6 +34,7 @@
 #include "core/config/engine.h"
 #include "core/io/image.h"
 #include "core/io/logger.h"
+#include "core/os/time_enums.h"
 #include "core/string/ustring.h"
 #include "core/templates/list.h"
 #include "core/templates/vector.h"
@@ -184,33 +185,6 @@ public:
 	virtual MainLoop *get_main_loop() const = 0;
 
 	virtual void yield();
-
-	enum Weekday : uint8_t {
-		WEEKDAY_SUNDAY,
-		WEEKDAY_MONDAY,
-		WEEKDAY_TUESDAY,
-		WEEKDAY_WEDNESDAY,
-		WEEKDAY_THURSDAY,
-		WEEKDAY_FRIDAY,
-		WEEKDAY_SATURDAY,
-	};
-
-	enum Month : uint8_t {
-		/// Start at 1 to follow Windows SYSTEMTIME structure
-		/// https://msdn.microsoft.com/en-us/library/windows/desktop/ms724950(v=vs.85).aspx
-		MONTH_JANUARY = 1,
-		MONTH_FEBRUARY,
-		MONTH_MARCH,
-		MONTH_APRIL,
-		MONTH_MAY,
-		MONTH_JUNE,
-		MONTH_JULY,
-		MONTH_AUGUST,
-		MONTH_SEPTEMBER,
-		MONTH_OCTOBER,
-		MONTH_NOVEMBER,
-		MONTH_DECEMBER,
-	};
 
 	struct DateTime {
 		int64_t year;
